castle/client/requirements.txt
PyOpenGL
pygame
aruco